+++ components.d.ts
@@ -31,6 +31,9 @@ declare module '@vue/runtime-core' {
     BaseInput: typeof import('./src/components/BaseInput.vue')['default']
     BaseMessage: typeof import('./src/components/BaseMessage.vue')['default']
     BaseQuestionAll: typeof import('./src/components/BaseQuestionAll.vue')['default']
+    BaseQuestionFormAll: typeof import('./src/components/BaseQuestionFormAll.vue')['default']
+    BaseQuestionFormScore: typeof import('./src/components/BaseQuestionFormScore.vue')['default']
+    BaseQuestionFormText: typeof import('./src/components/BaseQuestionFormText.vue')['default']
     BaseQuestionScore: typeof import('./src/components/BaseQuestionScore.vue')['default']
     BaseQuestionText: typeof import('./src/components/BaseQuestionText.vue')['default']
     BaseQuestionwll: typeof import('./src/components/BaseQuestionwll.vue')['default']
